Factor data feeds endpoints into new sub-mod
